.github/skills/artifacts/scripts/requirements.txt
# Runtime extraction stack for artifacts and artifacts-input-2-data pipelines
markitdown
Pillow
pillow-heif
pytesseract
faster-whisper
PyYAML
